requirements.txt
discord.py
pyyaml
asyncio
disputils
simplejson
pysqlite3
openai
requests